test(connlib): split out `Ref{Client,Gateway}`
